Merge

+++ PP2GameDevelopment/Assets/Scripts/collectibles/ICollectible.cs
@@ -21,7 +21,7 @@ public interface ICollectible
     //Addition to max Speed
     public void PlayerSpeedUp(float speed, float sprint);
     //Adding & Increasing crit chance
-    public void CritChanceUp(float chance, float dam);
+    public void CritChanceUp(int chance, float dam);
     //Adding dash uses and distance
     public void DashUp(int uses, float time);
